Fixes lots of serialization issues

+++ Projects/UOContent/Items/Resources/Blacksmithing/Ore.cs
@@ -43,8 +43,8 @@ namespace Server.Items
                 case 1:
                     {
                         // Use this line instead if you are getting world loading issues
-                        // _resource = (CraftResource)reader.ReadByte();
-                        _resource = (CraftResource)reader.ReadInt();
+                        _resource = (CraftResource)reader.ReadByte();
+                        // _resource = (CraftResource)reader.ReadInt();
                         break;
                     }
                 case 0:
